CalculateArea done

+++ csharp-basics/exercises/Arithmetic/CalculateArea/Geometry.cs
@@ -8,34 +8,37 @@ namespace CalculateArea
 {
     public class Geometry
     {
-        public static decimal AreaOfCircle(decimal radius)
+        public static double AreaOfCircle(decimal radius)
         {
             if (radius<0)
             {
                 Console.WriteLine("Negative value entered Area set to 0");
                 return 0;
             }
-            return (decimal)Math.PI*radius*2m;
+            return Math.PI*(double)radius*2d;
         }
 
-        public static decimal AreaOfRectangle(decimal length, decimal width)
+        public static double AreaOfRectangle(decimal length, decimal width)
         {
+
             if (length < 0 || width <0)
             {
                 Console.WriteLine("Negative value entered Area set to 0");
                 return 0;
             }
-            return length*width;
+            
+            return (double)(length * width);
         }
 
-        public static decimal AreaOfTriangle(decimal ground, decimal h)
+        public static double AreaOfTriangle(decimal ground, decimal h)
         {
             if (ground < 0 || h<0)
             {
                 Console.WriteLine("Negative value entered Area set to 0");
                 return 0;
             }
-            return ground*h*0.5m;
+            
+            return (double)(ground*h*0.5m);
         }
     }
 }
